process: make SIGPIPE an always failure
This error code does not exist on Windows, so any code that relied on
this was broken (would fail this test) on that platform already. In most
cases, it was actually a serious error too, such as `cp -v` or `tar -v`,
which will indeed fail midway and give corrupt results if you do not read
their output to completion!